Refactor to use pytest.
