+++ load_heka_python/test/test_load_heka.py
@@ -1,406 +1,131 @@
-import warnings
-import numpy as np
-from load_heka_python.load_heka import LoadHeka
+import sys
+sys.path.append(r"C:\fMRIData\git-repo\load-heka-python")
+
+from .runners import heka_reader_tester
 from os.path import join
+import pytest
 
-def test_heka_reader(base_path, version, group_series_to_test, dp_thr=1e-6, assert_time=True, include_stim_protocol=True):
+
+class TestFiles:
     """
-    Test the data read my LoadHeka matches the data when loading into Patchmaster. Data must be exported from Patchmaster
-    using 'Export Sweep' as ascii for comparison with LoadHeka. See the main README.md (section 5) on details
-    on the file organisation and Export sweep options expected by this function.
-
-    Note that all tests are conducted with units scaled to pA, mV, s.
-
-    This test function is a little unusual as by default it does not assert to an exact match between the data and Patchmaster.
-    For reasons that are not currently clear (have emailed HEKA) the Im and Vm data does not match the exported HEKA sweep
-    perfectly, but usually to around ~5 d.p. This varies a little between series. While this precision is sufficient, it would
-    be nice if they matched perfectly, and we could test with assert to 7 d.p. (similar to time and stimulus protocol, which are
-    not effected by this issue). Because of this, options to print the average decimal place match, or the % of samples
-    that match to dp_thr decimal places is provided.
-
-    base_path - path to the test directory (see README.md)
-    version - name you have given to this test file (see README.md)
-    group_series_to_test - list of group numbers (not index) and series to test (in str format)
-                           e.g. to test the third series of the second group, and the fourth series of the third group,
-                           = [["2", "3"], ["3", "4"]]
-
-    dp_thr - decimal place threshold, used as the number of decimal places to test matches.
-
-    assert_time - Tests are usually printed so the results for all series are run. However, if assert_time is True, an exception will be
-                  raised if data and Patchmaster do not match to pre-set cutoff values (see TestSeries().handle_assert() for details.
+    run as `pytest . -s` to see running output, as each individual
+    test can take some time to run through all series.
     """
-    print("Testing " + version + "----------------------------------------------------------------------------------------------------------------\n")
+    @pytest.fixture
+    def base_path(self):
+        return r"C:\fMRIData\git-repo\load-heka-python\ee_test_files\stim_extraction"
 
-    heka_file = LoadHeka(join(base_path, version, version + ".dat"),
-                         only_load_header=True)
+    def test_f1_v2x90_2(self, base_path):
 
-    for group_num, series_num in group_series_to_test:
-        print(f"group: {group_num} series: {series_num}")
-        filename = version + "_group-{0}_series-{1}.asc".format(group_num,
-                                                                series_num)
+        version = "f1_v2x90.2"
+        group_num_series_num_to_test = [["1", "1"], ["1", "18"], ["5", "6"],
+                                        ["7", "3"], ["8", "16"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-04)
 
-        raw_heka = read_heka_test_ascii(join(base_path, version, filename))
+    def test_f2_v2x90_2(self, base_path):
 
-        supress_time = supress_stim = False
-        for im_or_vm in ["Im", "Vm"]:
+        version = "f2_v2x90.2"
+        group_num_series_num_to_test = [["1", "4"], ["1", "1"], ["1", "2"], ["1", "3"], ["1", "4"], ["1", "5"],
+                                         ["1", "6"], ["1", "7"], ["1", "8"], ["1", "9"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-02)
 
-            load_heka = heka_file.get_series_data(im_or_vm,
-                                                  int(group_num) - 1,
-                                                  int(series_num) - 1,
-                                                  include_stim_protocol=include_stim_protocol,
-                                                  fill_with_mean=True)
-            test = TestSeries(raw_heka,
-                              load_heka,
-                              group_num,
-                              series_num,
-                              dp_thr,
-                              im_or_vm,
-                              assert_time,
-                              supress_time,
-                              supress_stim,
-                              )
+    def test_f3_v2x90_2(self, base_path):
 
-            if test.stim_tested_flag:
-                supress_stim = True
-            if test.time_tested_flag:
-                supress_time = True
+        version = "f3_v2x90.2"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"],  ["1", "3"], ["1", "4"], ["1", "5"],
+                                        ["1", "6"], ["1", "7"], ["1", "8"], ["1", "9"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-03)
 
-        print("\n")
 
-class TestSeries:
-    def __init__(self, raw_heka, load_heka, group_num, series_num, dp_thr, im_or_vm, assert_time, supress_time, supress_stim):
+    def test_f4_v2x90_5(self, base_path):
+
+        version = "f4_v2x90.5"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"], ["1", "3"], ["1", "4"], ["1", "5"], ["1", "14"], ["1", "16"],
+                                         ["1", "17"], ["1", "18"], ["1", "19"], ["1", "20"], ["1", "22"], ["1", "23"], ["1", "24"],
+                                         ["1", "25"], ["1", "26"], ["1", "27"], ["1", "28"], ["1", "29"], ["1", "30"], ["1", "31"],
+                                         ["1", "32"], ["1", "33"], ["1", "34"], ["1", "35"], ["1", "36"],
+
+                                         ["2", "1"], ["2", "2"], ["2", "3"], ["2", "4"], ["2", "5"], ["2", "6"], ["2", "7"], ["2", "9"], ["2", "12"], ["2", "14"],
+                                         ["2", "15"], ["2", "16"], ["2", "18"], ["2", "21"], ["2", "24"], ["2", "25"], ["2", "26"], ["2", "27"], ["2", "28"], ["2", "29"],
+                                         ["2", "30"], ["2", "31"], ["2", "32"], ["2", "33"], ["2", "34"], ["2", "35"], ["2", "36"], ["2", "37"], ["2", "38"], ["2", "39"],
+                                         ["2", "40"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-03)
+
+
+    def test_f5_v2x90_5(self, base_path):
+        version = "f5_v2x90.5"
+        group_num_series_num_to_test = [["1", "1"], ["1", "4"], ["1", "5"], ["1", "6"],  # ["3", "3"], series has records with three channels
+                                         ["1", "7"], ["3", "9"], ["5", "6"], ["5", "8"],
+                                         ["5", "10"],  ["6", "2"], ["6", "14"], ["6", "20"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-03)
+
+
+    def f6_1_2_0_build_1469(self, base_path):
+        version = "f6_1.2.0_[Build 1469]"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"], ["1", "3"], ["1", "4"],  # ["1", "6"] series has records with three channels
+                                         ["1", "5"], ["1", "7"], ["1", "8"],
+                                         ["1", "9"], ["1", "10"], ["1", "11"], ["1", "12"], ["1", "13"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test,  assert_time=False, dp_thr=1e-02)
+
+    def f7_1_2_0_build_1469(self, base_path):
+
+        version = "f7_1.2.0_[Build 1469]"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"], ["1", "3"], ["1", "4"], ["1", "5"],  # ["1", "8"], series has records with three channels
+                                         ["1", "6"], ["1", "7"], ["1", "9"], ["1", "10"],
+                                         ["1", "11"], ["1", "12"], ["1", "13"], ["1", "14"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-03)
+
+    def test_f8_v2x65(self, base_path):
+        """Stimulus reconstruction for versions before 2x90 is not supported"""
+        version = "f8_v2x65"
+        group_num_series_num_to_test = [["1", "6"], ["1", "13"], ["3", "1"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-05, include_stim_protocol=False)
+
+    def test_f9_v2x65(self, base_path):
+        """Stimulus reconstruction for versions before 2x90 is not supported"""
+        version = "f9_v2x65"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-04, include_stim_protocol=False)
+
+    def test_f10_v2x91(self, base_path):
+        version = "f10_v2x91"
+        group_num_series_num_to_test = [["1", "1"], ["1", "3"], ["1", "5"], ["1", "9"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-04)
+
+    def test_f11_v2x90_3(self, base_path):
         """
-        Class to test the Data (Im or Vm), reconstructed stimulus and time against exported Patchmaster data. Data is converted to pA
-        and mV. For a reason that is not clear, the output of the Im and Vm data does not match HEKA perfectly, but to around 5 decimal
-        places on average. This precision is sufficient but we have emailed HEKA to check if any additional settings are causing
-        this slight difference. See test_heka_reader() for details.
+        This is an annoying file that is offset by 0.005 s in the software (x_start = 0.005) but
+        cannot export it in such a way that it is relative to the sweep but maintains
+        the 0.05 s offset - the software is correcting it somewhere but not making the
+        option to turn off correction available (at least that I could find).
+        This is not a severe issue for our tests as the time is read correctly, just
+        the offset is not automatically removed (as it is in HEKA).
+        Nonetheless Vm is loading well so keep here as a successful test.
 
-        NOTES:
-            For the HEKA exported data, the number of samples in each stimulus segment does not match the HEKA Gui. For example,
-            e.g. in Heka test file f3, the stimulus segments should have 5000, 50000, 10000 samples.
-            However the stimulus output actually as 5001, 50000, 4999 samples.
-
-            Here the HEKA data are  shifted back 1 idx and extended 1 index so that the samples are correct (i.e. match the HEKA GUI)
-             - this is done under the assumption that the HEKA GUI is correct.
-
-            For the ramp, f2 group-1 series-8 should be 20,000 samples, 20,0000 samples, 20,000 samples. The first step of the ramp is sample 20,002,
-            the last sample is 220,001 (this is exactly 200,000 samples). Thus the ramp does not start at 0 but at 0 + one step.
-
-            Data and time must be tested per-sweep because they can be different lengths but the reconstructured stimulis is always the same num samples
-            per sweep
+        There is also a problem with the stimulus, which is all zeros in the
+        file but was loading with array dimension longer than the time of the recording
+        in load_heka, so is loaded as False. This is most likely due to the user manually stopping
+        the recording before it has finished.
         """
-        self.raw_heka = raw_heka
-        self.load_heka = load_heka
-        self.group_num = group_num
-        self.series_num = series_num
-        self.im_or_vm = im_or_vm
-        self.assert_time = assert_time
-        self.dp_thr = dp_thr
-        self.stim_tested_flag = False
-        self.time_tested_flag = False
+        version = "f11_v2x90.3"
+        group_num_series_num_to_test = [["1", "1"], ["1", "2"], ["1", "4"], ["1", "11"], ["1", "12"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-04, assert_time=False, include_stim_protocol=False)
 
-        self.test_data()
-
-        if not supress_time:
-            self.test_time()
-
-        if not supress_stim:
-            self.test_reconstructed_stimulus()
-
-    def test_data(self):
-        all_assert = []
-        if np.any(self.load_heka["data"]):
-
-            for raw_heka_sweep, load_heka_sweep in zip(self.raw_heka[self.im_or_vm],
-                                                       self.load_heka["data"]):
-
-                raw = self.clean_data(raw_heka_sweep, expected_len=len(load_heka_sweep))
-                loaded = self.clean_data(load_heka_sweep)
-
-                all_assert.append(np.allclose(raw, loaded, rtol=0, atol=self.dp_thr))
-
-            assert all(all_assert), "group {0} series {1} does not match Patchmaster to {2} decimal places".format(self.group_num,
-                                                                                                                   self.series_num,
-                                                                                                                   self.printable_dp())
-            print("group {0} series {1} matches Patchmaster to exactly {2} decimal places".format(self.group_num, self.series_num, self.printable_dp()))
-
-# Test Time ------------------------------------------------------------------------------------------------------------------------------------------
-
-    def test_time(self):
+    def test_f12_v2x90_4(self, base_path):
         """
+        The same issue with som stimulus reconstruction as described above is observed on thi
+        test file. However, most stimulus are reconstructed successfully, further suggesting it
+        is a problem with the user manually stopping the recording too quickly.
         """
-        if not np.any(self.load_heka["time"]):
-            return
+        version = "f12_v2x90.4"
+        group_num_series_num_to_test = [["1", "1"], ["1", "3"], ["1", "5"], ["3", "3"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test,  dp_thr=1e-03,  include_stim_protocol=False)
 
-        time_sweeps_correct = []
-        for raw_sweep_time, load_heka_sweep_time in zip(self.raw_heka["time_" + self.im_or_vm],
-                                                        self.load_heka["time"]):
-
-            sweep_correct = np.allclose(self.clean_time(raw_sweep_time, expected_len=len(load_heka_sweep_time)),
-                                        load_heka_sweep_time,
-                                        atol=1e-10, rtol=0)
-            time_sweeps_correct.append(sweep_correct)
-
-        if all(time_sweeps_correct):
-            print("time correct for group {0} series {1}".format(self.group_num,
-                                                                 self.series_num))
-        else:
-            print("Time INCORRECT for group {0} series {1}".format(self.group_num,
-                                                                   self.series_num))
-        self.time_tested_flag = True
-
-        if self.assert_time:
-            assert all(time_sweeps_correct), "Time does not match Patchmaster to 10 decimal places"
-
-# Test Stimululus ------------------------------------------------------------------------------------------------------------------------------------
-
-    def test_reconstructed_stimulus(self):
+    def test_f13_v2x92(self, base_path):
         """
+        This is the only file that contains interleaved records.
+        Some stimulus are NaN in this file, so not incldued.
         """
-        if not self.load_heka["stim"]:
-            return
-
-        raw_stim = self.get_raw_stim()
-        if raw_stim is False:
-            return
-
-        load_heka_stim = self.conv(self.load_heka["stim"]["data"], self.load_heka["stim"]["units"])  # stim output is A and V for consistency
-
-        stim_is_close = np.allclose(raw_stim, load_heka_stim, atol=1e-07, rtol=0)
-
-        if stim_is_close:
-            print("stimulus was correctly reconstructed for group {0} series {1} {2}".format(self.group_num,
-                                                                                             self.series_num,
-                                                                                             self.im_or_vm))
-        else:
-            print("stimulus was NOT correctly reconstructed for group {0} series {1} {2}".format(self.group_num,
-                                                                                                 self.series_num,
-                                                                                                 self.im_or_vm))
-        self.stim_tested_flag = True
-
-        assert stim_is_close, "Stimulus reconstruction does not match Patchmaster to 7 decimal places"
-
-
-    def get_raw_stim(self):
-        """
-        raw stim is output from ascii script as ragged array of numpy, so necessary to check they are all the same size.
-        We also need to fix the issue with HEKA export output, see __Init__ doc for details
-        """
-        raw_heka_output = self.raw_heka["stim_" + self.im_or_vm]
-
-        if not self.all_sweeps_are_same_num_samples(raw_heka_output):
-            warnings.warn("stimulus not tested, the record sizes are not equal for all sweeps")
-            return False
-
-        stim_data = self.clean_raw_stim(raw_heka_output)
-        stim_data = stim_data[:, 1:]
-        stim_data = np.append(stim_data, np.atleast_2d(stim_data[:, -1]).T, axis=1)
-
-        return stim_data
-
-# Utils ----------------------------------------------------------------------------------------------------------------------------------------------
-
-    def clean_data(self, data, expected_len=None):
-        """
-        """
-        if expected_len:
-            if len(data) != expected_len:
-                data = self.pad_short_data_with_mean(data, expected_len)
-
-        data = self.conv(data,
-                         self.im_or_vm)
-        return data
-
-    def clean_time(self, data, expected_len=None):
-
-        if expected_len:
-            if len(data) != expected_len:
-                data = self.pad_short_time_with_extra_time(data, expected_len)
-
-        return data
-
-    @staticmethod
-    def pad_short_time_with_extra_time(data, expected_len):
-        """
-        HEKA export data is not full, is a sweep is cut short it will be missing the end. However HekaLoad will always
-        export full time. So pad heka export data here to match
-        """
-        new_data = np.full(expected_len, np.nan)
-        len_data = len(data)
-        len_empty_data = expected_len - len_data
-
-        new_data[0:len_data] = data
-        ts = data[1] - data[0]
-
-        pad_data = np.arange(1, len_empty_data + 1) * ts + data[-1]
-        new_data[len_data:] = pad_data
-
-        return new_data
-
-    @staticmethod
-    def pad_short_data_with_mean(data, expected_len):
-
-        len_short_data = len(data)
-
-        new_data = np.full(expected_len, np.nan)
-        new_data[0:len_short_data] = data
-        new_data[len_short_data:] = np.mean(data)
-
-        return new_data
-
-    def clean_raw_stim(self, raw_heka_output):
-        """
-        Raw stim is output from ascii script as a list of np arrays (sweeps)
-
-        For some reason when stim doesn't exist heka export as NaN but LoadHeka will export as holding current
-        """
-        stim_data = np.array(raw_heka_output)
-        if np.isnan(stim_data).all():
-            stim_data = np.zeros(np.shape(stim_data))
-
-        conv_stim_data = self.conv(stim_data,
-                                   self.load_heka["stim"]["units"])
-        return conv_stim_data
-
-    @staticmethod
-    def conv(data, units):
-        """
-        HEKA data is always saved in A or V (this is tested with assert on import). Here we want it in pA or mV
-        for both consistency with EE and easier testing.
-        """
-        if units in ["Im", "A"]:
-            factor = 1000000000000
-        elif units in ["Vm", "V"]:
-            factor = 1000
-        elif units == "mV":
-            factor = 1
-
-        return data * factor
-
-    @staticmethod
-    def all_sweeps_are_same_num_samples(raw_heka_output):
-        return len(np.unique([len(sweep) for sweep in raw_heka_output])) == 1
-
-    def printable_dp(self):
-        return abs(int(f'{self.dp_thr:e}'.split('e')[-1]))
-
-# ----------------------------------------------------------------------------------------------------------------------------------------------------
-# Read Patchmaster Export Sweep ASCII
-# ----------------------------------------------------------------------------------------------------------------------------------------------------
-
-def read_heka_test_ascii(full_filepath):
-    """
-    Read HEKA Export Sweet ASCII data into a dictionary of data.
-
-    It will load a exported series into a dictionary array including data (e.g. Vm), time (e.g. time_Vm) and stimulus (e.g. stim_Vm)
-    for Vm and Im. These are saved as list of numpy array (because each sweep num samples is not garenteed to be the same)
-    in the relevant dict fields.
-
-    NOTE: this function expects Exported series in the typical format, with 3 lines before the first data line. In some cases, HEKA will
-          export with 4 lines. In this case, at present the text file must be manually edited to move / delete this 4th line.
-
-    INPUT: full_filepath - full path to the ascii
-
-    OUTPUT:
-        results dictionary with keys: Im, time_Im, stim_Im, Vm, time_Vm, stim_Vm.
-    """
-    with open(full_filepath) as fh:
-
-        results = make_empty_param_dict()
-        sweep = make_empty_param_dict()
-
-        fh.seek(0)
-        i = -1
-        while True:
-
-            line = fh.readline()
-            i += 1
-
-            if "Time" in line:
-                cols = line
-                idx_col_present = True if "Index" in cols else False
-                stim_present = True if "Stimulus" in cols else False
-                continue
-
-            elif "Series" in line:
-                continue
-
-            elif line == "\n":
-                continue
-
-            elif line == "":
-                break
-
-            elif "Sweep" in line:
-                if i > 3:  # ignore first 'Sweep' because it comes at the start of the data block
-
-                    update_results_with_sweep(results, sweep)
-
-                    sweep = make_empty_param_dict()
-            else:
-                idx_col_offset = 1 if idx_col_present else 0
-                if "[A]" in cols.split(",")[1 + idx_col_offset]:
-                    save_line_to_sweep_dict(line, sweep, first="Im", second="Vm", idx_col_present=idx_col_present, stim_present=stim_present)
-
-                elif "[V]" in cols.split(",")[1 + idx_col_offset]:
-                    save_line_to_sweep_dict(line, sweep, first="Vm", second="Im", idx_col_present=idx_col_present, stim_present=stim_present)
-                else:
-                    raise Exception("ascii col not recognised")
-
-        update_results_with_sweep(results, sweep)  # get final sweep_idx
-
-        return results
-
-def save_line_to_sweep_dict(line, sweep, first, second, idx_col_present, stim_present):
-    """
-    The output ascii will have either 3 or 6 columns, depending on whether Im and Vm or just Im or Vm records are present.
-    They can be in either order (e.g. first 3 cols Vm)
-
-    Here save the Im and Vm to the approrpiate fields in the dict accounting for whether Im or Vm is first
-    """
-    split_line = line.strip().split(",")
-    handle_line(sweep, split_line,
-                start_idx=0, stop_idx=3, label=first,
-                idx_col_present=idx_col_present, stim_present=stim_present)
-
-    if len(split_line) > 3:
-        handle_line(sweep, split_line,
-                    start_idx=3, stop_idx=6, label=second,
-                    idx_col_present=idx_col_present, stim_present=stim_present)
-
-    else:
-        sweep["time_" + second].append(np.nan)
-        sweep[second].append(np.nan)
-        sweep["stim_" + second].append(np.nan)
-
-
-def handle_line(sweep, split_line, start_idx, stop_idx, label, idx_col_present, stim_present):
-    """
-    For file versions >2x90, there is no index column and we can open in Patchamsdter so export with stimulus.
-    For older versions, there is a leading index coloumn and stimulus does not always export.
-    This handles both cases here, indexing into the read line appropriately.
-    """
-    if idx_col_present and not stim_present:
-        idx_offset = 1
-        stim_offset = 0 if stop_idx == 3 else -1
-
-        line_time, line_data = split_line[start_idx + stim_offset + idx_offset: stop_idx + stim_offset]
-        line_stim = np.nan
-
-    elif not idx_col_present and stim_present:
-        line_time, line_data, line_stim = split_line[start_idx:stop_idx]
-
-    sweep["time_" + label].append(float(line_time))
-    sweep[label].append(float(line_data))
-    sweep["stim_" + label].append(float(line_stim))
-
-def update_results_with_sweep(results, sweep):
-    for key, value in sweep.items():
-        results[key].append(np.array(value))
-
-def make_empty_param_dict():
-    dict_ = {}
-    for key in ["Im", "time_Im", "Vm", "time_Vm", "stim_Im", "stim_Vm"]:
-        dict_[key] = []
-
-    return dict_
+        version = "f13_v2x92"
+        group_num_series_num_to_test = [["1", "4"], ["2", "2"], ["3", "4"], ["5", "4"], ["6", "2"]]
+        heka_reader_tester(base_path, version, group_num_series_num_to_test, dp_thr=1e-04, include_stim_protocol=False)
